Reduce the number of requests made by S3Files
As written, the S3File implementations issue one request to AWS per call
to Read. When the ParquetReader issues repeated small Read calls, this
results in a large number of S3 GetObject requests, which causes bad
performance and high AWS costs.

This commit reimplements the reader logic for the two versions of S3File
in the following way:

When reading, if there is no open reader into the S3 object, it issues a
GetObject request with a large range and stores the io.ReadCloser into
the body in the S3File object. This io.ReadCloser will be serve any
incoming Read calls. When the ParquetReader Seeks to a new location in
the file or an error occurs while reading the file, the io.ReadCloser is
closed and discarded, and the next call to Read will issue a new
request.

A new configuration parameter "MinRequestSize" is added to control the
amount of over-fetching. If not specified, we pick a generous default.

A new constructor NewS3FileReaderWithParams is added, which accepts a
new S3FileReaderParams that contains all the fields used to construct
an S3File for reading. This should help mitigate the problem of
proliferating constructors for different sets of configuration
parameters - the caller sets the parameters they care about and ignores
the ones they don't, and the constructor handles default behavior.