fix: Speed up GCS reader
The previous implementation closed the file reader on each call to
`Read` which made the operation very expensive, especially on large
files when only reading 4k/call.

Instead of re-implementing a reader with support for `io.Seek`, this PR
changes the implementation to use an existing reader for storage
objects that supports `Seek`.

Comparing this change with a parquet file of 200LOC with 15 columns
around 80M the previous implementation took ~6m30s and the current
implementation takes ~18s.

In addition to that, this PR simplifies what we need to keep on the
struct to be able to implement `Open` and `Create`.

This PR also does some hygiene to use more idiomatic Go:
* Don't use `self` or `this` for receiver name
* Use capital `ID` for identifiers
* Add compile time check that we implement `sources.ParquetFile`
* Rename `GcsFile` to `File` to make less stuttering (`gcs.File`)